Ok I've made quite a few changes
