--- data_collection/README.md
@@ -1,2 +1,0 @@
-# this is the data collection folder, where all the init data collection was done
-## see scrape.ipynb
